requirements.txt
requests>=2.28.0
python-dotenv>=0.20.0
fastapi>=0.95.0
uvicorn>=0.21.0 